+++ cli/src/init/service.rs
@@ -528,7 +528,7 @@ pub(crate) fn generate_service_package_json(
                 types_pino: None,
                 types_ioredis: None,
                 pino: None,
-                ioredis: Some(IOREDIS_VERSION.to_string()),
+                ioredis: None,
                 additional_deps: HashMap::new(),
             }
         }),