+++ blueprint/implementations/billing/stripe/CHANGELOG.md
@@ -1,4 +1,13 @@
-# @forklaunch/implementation-billing-base
+# @forklaunch/implementation-billing-stripe
+
+## 0.5.9
+
+### Patch Changes
+
+- dependency upgrade and minor metadata fixes
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@0.8.9
+  - @forklaunch/interfaces-billing@0.8.9
 
 ## 0.5.8