+++ blueprint/implementations/iam/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-iam-base
 
+## 0.8.9
+
+### Patch Changes
+
+- dependency upgrade and minor metadata fixes
+- Updated dependencies
+  - @forklaunch/interfaces-iam@0.8.9
+
 ## 0.8.8
 
 ### Patch Changes