+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 0.8.9
+
+### Patch Changes
+
+- dependency upgrade and minor metadata fixes
+
 ## 0.8.8
 
 ### Patch Changes