+++ blueprint/interfaces/worker/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @forklaunch/interfaces-billing
+# @forklaunch/interfaces-worker
+
+## 0.7.9
+
+### Patch Changes
+
+- dependency upgrade and minor metadata fixes
 
 ## 0.7.8