+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 0.11.2
+
+### Patch Changes
+
+- Package deps version bump
+- Updated dependencies
+  - @forklaunch/validator@0.10.26
+  - @forklaunch/common@0.6.26
+  - @forklaunch/core@0.17.2
+  - @forklaunch/ws@0.1.5
+
 ## 0.11.1
 
 ### Patch Changes