+++ framework/testing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/testing
 
+## 0.0.15
+
+### Patch Changes
+
+- Package deps version bump
+
 ## 0.0.14
 
 ### Patch Changes